@@ -1,57 +1,38 @@
 {
-  "name": "react-landing-page-2021",
-  "version": "1.0.0",
+  "name": "web-nawara-p",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "build-stats": "cross-env ANALYZE=true npm run build",
-    "export": "next export",
-    "build-prod": "run-s clean build export",
-    "clean": "rimraf .next out",
-    "lint": "next lint",
-    "build-types": "tsc --noEmit --pretty",
-    "prepare": "husky install"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@headlessui/react": "^1.4.1",
-    "@heroicons/react": "^1.0.5",
-    "framer-motion": "^5.0.0",
-    "next": "^12.0.0",
-    "next-seo": "^4.28.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scroll": "^1.8.4",
-    "smooth-scroll": "^16.1.3",
-    "styled-jsx-plugin-postcss": "^4.0.1"
+    "@react-hook/media-query": "^1.1.1",
+    "animate.css": "^4.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-intersection-observer": "^9.8.1",
+    "react-router-dom": "^6.22.1",
+    "react-scroll": "^1.9.0"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^12.0.0",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@typescript-eslint/eslint-plugin": "^4.32.0",
-    "@typescript-eslint/parser": "^4.32.0",
-    "autoprefixer": "^10.3.7",
-    "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-airbnb-typescript": "^14.0.1",
-    "eslint-config-next": "^12.0.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-unused-imports": "^1.1.5",
-    "husky": "^7.0.4",
-    "lint-staged": "^11.2.5",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.3.11",
-    "prettier": "^2.4.1",
-    "rimraf": "^3.0.2",
-    "tailwindcss": "^2.2.17",
-    "typescript": "^4.4.4"
-  },
-  "license": "ISC"
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "@types/react-scroll": "^1.8.10",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.17",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.1.4"
+  }
 }
