@@ -1,6 +1,6 @@
-// Please do not use the array form (like ['tailwindcss', 'postcss-preset-env'])
-// it will create an unexpected error: Invalid PostCSS Plugin found: [0]
-
-module.exports = {
-  plugins: { tailwindcss: {}, autoprefixer: {} },
-};
+export default {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}
