@@ -1,47 +1,17 @@
-module.exports = {
-  mode: 'jit',
-  purge: ['./src/**/*.{js,ts,jsx,tsx}'],
-  darkMode: false,
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
-    fontSize: {
-      xs: '0.75rem',
-      sm: '0.875rem',
-      base: '1rem',
-      lg: '1.125rem',
-      xl: '1.25rem',
-      '2xl': '1.5rem',
-      '3xl': '1.875rem',
-      '4xl': '2.25rem',
-      '5xl': '3rem',
-      '6xl': '4rem',
-    },
     extend: {
+      height: {
+        "calc-64": "calc(100vh - 64px)",
+      },
       colors: {
-        primary: '#179A9E',
-        secondary: '#0F7477',
-        tertiary: '#DD295F',
-        quaternary: '#1a2e35',
-        border: '#1a2e35',
-        background: '#ffffff',
-      },
-      animation: {
-        vote: 'vote 1s ease-in-out',
-      },
-      keyframes: {
-        vote: {
-          '0%, 100%': {
-            transform: 'rotate(0deg)',
-          },
-          '25%': {
-            transform: 'rotate(-30deg)',
-          },
-          '75%': {
-            transform: 'rotate(30deg)',
-          },
-        },
+        primary: "#BFEBE5",
+        secondary: "#1a2e35",
+        tertiary: "#179a9e",
       },
     },
   },
-  variants: {},
   plugins: [],
 };
