@@ -1,8 +1,8 @@
-import { useRef, FC, useEffect, useState } from 'react';
+import { useRef, FC, useEffect, useState } from "react";
 
-import { CanvasContext } from '../hooks/useCanvas';
-import useResponsiveSize from '../hooks/useResponsiveSize';
-import Wave from './Wave';
+import { CanvasContext } from "../hooks/useCanvas";
+import useResponsiveSize from "../hooks/useResponsiveSize";
+import Wave from "./Wave";
 
 const Canvas: FC = () => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
@@ -12,7 +12,7 @@ const Canvas: FC = () => {
   >();
 
   useEffect(() => {
-    const ctx = canvasRef?.current?.getContext('2d');
+    const ctx = canvasRef?.current?.getContext("2d");
     if (ctx) setContext(ctx);
   }, []);
 
