@@ -5,7 +5,7 @@ interface ItemListaProps {
 
 const ItemList: React.FC<ItemListaProps> = ({ id, texto }) => (
   <li>
-    <a href={`#${id}`} className="no-underline text-quaternary">
+    <a href={`#${id}`} className="no-underline text-tertiary">
       {texto}
     </a>
   </li>
@@ -16,37 +16,37 @@ interface ListaOrdenadaProps {
 }
 
 export const datosLista = [
-  { id: 'pregunta1', texto: '¿QUÉ INFORMACIÓN RECOPILAMOS?' },
-  { id: 'pregunta2', texto: '¿CÓMO PROCESAMOS TU INFORMACIÓN?' },
+  { id: "pregunta1", texto: "¿QUÉ INFORMACIÓN RECOPILAMOS?" },
+  { id: "pregunta2", texto: "¿CÓMO PROCESAMOS TU INFORMACIÓN?" },
   {
-    id: 'pregunta3',
-    texto: '¿CUÁNDO Y CON QUIÉN COMPARTIMOS TU INFORMACIÓN PERSONAL?',
+    id: "pregunta3",
+    texto: "¿CUÁNDO Y CON QUIÉN COMPARTIMOS TU INFORMACIÓN PERSONAL?",
   },
   {
-    id: 'pregunta4',
-    texto: '¿UTILIZAMOS COOKIES Y OTRAS TECNOLOGÍAS DE RASTREO?',
+    id: "pregunta4",
+    texto: "¿UTILIZAMOS COOKIES Y OTRAS TECNOLOGÍAS DE RASTREO?",
   },
   {
-    id: 'pregunta5',
-    texto: '¿CÓMO MANEJAMOS TUS INICIOS DE SESIÓN EN REDES SOCIALES?',
+    id: "pregunta5",
+    texto: "¿CÓMO MANEJAMOS TUS INICIOS DE SESIÓN EN REDES SOCIALES?",
   },
-  { id: 'pregunta6', texto: '¿CUÁNTO TIEMPO GUARDAMOS TU INFORMACIÓN?' },
-  { id: 'pregunta7', texto: '¿CÓMO MANTENEMOS SEGURA TU INFORMACIÓN?' },
-  { id: 'pregunta8', texto: '¿RECOLECTAMOS INFORMACIÓN DE MENORES?' },
-  { id: 'pregunta9', texto: '¿CUÁLES SON TUS DERECHOS DE PRIVACIDAD?' },
-  { id: 'pregunta10', texto: 'CONTROLES PARA FUNCIONES DE NO SEGUIMIENTO' },
+  { id: "pregunta6", texto: "¿CUÁNTO TIEMPO GUARDAMOS TU INFORMACIÓN?" },
+  { id: "pregunta7", texto: "¿CÓMO MANTENEMOS SEGURA TU INFORMACIÓN?" },
+  { id: "pregunta8", texto: "¿RECOLECTAMOS INFORMACIÓN DE MENORES?" },
+  { id: "pregunta9", texto: "¿CUÁLES SON TUS DERECHOS DE PRIVACIDAD?" },
+  { id: "pregunta10", texto: "CONTROLES PARA FUNCIONES DE NO SEGUIMIENTO" },
   {
-    id: 'pregunta11',
-    texto: '¿REALIZAMOS ACTUALIZACIONES A ESTA NOTIFICACIÓN?',
-  },
-  {
-    id: 'pregunta12',
-    texto: '¿CÓMO PUEDES CONTACTARNOS SOBRE ESTA NOTIFICACIÓN?',
+    id: "pregunta11",
+    texto: "¿REALIZAMOS ACTUALIZACIONES A ESTA NOTIFICACIÓN?",
   },
   {
-    id: 'pregunta13',
+    id: "pregunta12",
+    texto: "¿CÓMO PUEDES CONTACTARNOS SOBRE ESTA NOTIFICACIÓN?",
+  },
+  {
+    id: "pregunta13",
     texto:
-      '¿CÓMO PUEDES REVISAR, ACTUALIZAR O ELIMINAR LOS DATOS QUE RECOPILAMOS DE TI?',
+      "¿CÓMO PUEDES REVISAR, ACTUALIZAR O ELIMINAR LOS DATOS QUE RECOPILAMOS DE TI?",
   },
 ];
 
