@@ -1,8 +1,8 @@
-import { FC } from 'react';
+import { FC } from "react";
 
-import { useCanvasContext } from '../hooks/useCanvas';
-import useResponsiveSize from '../hooks/useResponsiveSize';
-import WaveObj from '../utils/wave';
+import { useCanvasContext } from "../hooks/useCanvas";
+import useResponsiveSize from "../hooks/useResponsiveSize";
+import WaveObj from "../utils/wave";
 
 const Wave: FC = () => {
   const { context } = useCanvasContext();
@@ -10,8 +10,8 @@ const Wave: FC = () => {
   const height = 600;
   let frequency = 0.013;
   const waves = {
-    frontWave: new WaveObj([0.0211, 0.028, 0.015], 'rgb(23, 154, 158, 0.1)'),
-    backWave: new WaveObj([0.0122, 0.018, 0.005], 'rgb(39, 158, 164, 0.1)'),
+    frontWave: new WaveObj([0.0211, 0.028, 0.015], "rgb(23, 154, 158, 0.1)"),
+    backWave: new WaveObj([0.0122, 0.018, 0.005], "rgb(39, 158, 164, 0.1)"),
   };
 
   const render = () => {
