@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from "react";
 
 const useResponsiveSize = () => {
   const [width, setWidth] = useState(0);
@@ -10,7 +10,7 @@ const useResponsiveSize = () => {
   }, [setWidth, setHeight]);
 
   useEffect(() => {
-    window.addEventListener('resize', setSizes);
+    window.addEventListener("resize", setSizes);
     setSizes();
   }, [setSizes]);
 
