@@ -1,54 +1,27 @@
-import React from 'react';
+import About from "../components/About";
+import Features from "../components/Features";
+import Product from "../components/Product";
+import Whatsapp from "../components/Whatsapp";
+import LandingPage from "../components/LandingPage";
+import Canvas from "../components/Canvas";
+import Navbar from "../components/Navbar";
 
-import About from '../components/About';
-import Analytics from '../components/Analytics';
-import Canvas from '../components/Canvas';
-import Features from '../components/Features';
-import Header from '../components/Header';
-import LazyShow from '../components/LazyShow';
-import MainHero from '../components/MainHero';
-import MainHeroImage from '../components/MainHeroImage';
-import Product from '../components/Product';
-
-const App = () => {
+function Index() {
   return (
-    <div className={`bg-background grid gap-y-16 overflow-hidden`}>
-      <div className={`relative bg-background`}>
-        <div className="max-w-7xl mx-auto">
-          <div
-            className={`relative z-10 pb-8 bg-background sm:pb-16 md:pb-20 lg:max-w-2xl lg:w-full lg:pb-28 xl:pb-32`}
-          >
-            <Header />
-            <MainHero />
-          </div>
-        </div>
-        <MainHeroImage />
-      </div>
+    <div className="max-w-screen">
+      <Navbar />
+      <LandingPage />
       <Canvas />
-      <LazyShow>
-        <>
-          <Product template="product" />
-          <Canvas />
-        </>
-      </LazyShow>
-      <LazyShow>
-        <>
-          <Features />
-          <Canvas />
-        </>
-      </LazyShow>
-      <LazyShow>
-        <Product template="howWorks" />
-      </LazyShow>
-      <LazyShow>
-        <>
-          <Canvas />
-          <About />
-        </>
-      </LazyShow>
-      <Analytics />
+      <Product template="product" />
+      <Canvas />
+      <Features />
+      <Canvas />
+      <Product template="howWorks" />
+      <Canvas />
+      <About />
+      <Whatsapp />
     </div>
   );
-};
+}
 
-export default App;
+export default Index;
